change to lasted construction, and modified some things not important
