+++ .idea/encodings.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="Encoding">
+    <file url="file://$PROJECT_DIR$/output_file/2019-6-4/station_geo.txt" charset="GBK" />
+    <file url="file://$PROJECT_DIR$/output_file/2019-6-4/station_geo.txt" charset="GBK" />
+  </component>
+</project>
