+++ .idea/inspectionProfiles/Project_Default.xml
@@ -1,0 +1,53 @@
+<component name="InspectionProjectProfileManager">
+  <profile version="1.0">
+    <option name="myName" value="Project Default" />
+    <inspection_tool class="PyPackageRequirementsInspection" enabled="true" level="WARNING" enabled_by_default="true">
+      <option name="ignoredPackages">
+        <value>
+          <list size="31">
+            <item index="0" class="java.lang.String" itemvalue="mlxtend" />
+            <item index="1" class="java.lang.String" itemvalue="joblib" />
+            <item index="2" class="java.lang.String" itemvalue="threadpoolctl" />
+            <item index="3" class="java.lang.String" itemvalue="scikit-learn" />
+            <item index="4" class="java.lang.String" itemvalue="python-dateutil" />
+            <item index="5" class="java.lang.String" itemvalue="numpy" />
+            <item index="6" class="java.lang.String" itemvalue="requests" />
+            <item index="7" class="java.lang.String" itemvalue="filelock" />
+            <item index="8" class="java.lang.String" itemvalue="lxml" />
+            <item index="9" class="java.lang.String" itemvalue="soupsieve" />
+            <item index="10" class="java.lang.String" itemvalue="PyQRCode" />
+            <item index="11" class="java.lang.String" itemvalue="urllib3" />
+            <item index="12" class="java.lang.String" itemvalue="zmail" />
+            <item index="13" class="java.lang.String" itemvalue="pyparsing" />
+            <item index="14" class="java.lang.String" itemvalue="imbox" />
+            <item index="15" class="java.lang.String" itemvalue="scipy" />
+            <item index="16" class="java.lang.String" itemvalue="pypng" />
+            <item index="17" class="java.lang.String" itemvalue="six" />
+            <item index="18" class="java.lang.String" itemvalue="kiwisolver" />
+            <item index="19" class="java.lang.String" itemvalue="packaging" />
+            <item index="20" class="java.lang.String" itemvalue="pandas" />
+            <item index="21" class="java.lang.String" itemvalue="fonttools" />
+            <item index="22" class="java.lang.String" itemvalue="itchat-uos" />
+            <item index="23" class="java.lang.String" itemvalue="platformdirs" />
+            <item index="24" class="java.lang.String" itemvalue="matplotlib" />
+            <item index="25" class="java.lang.String" itemvalue="virtualenv" />
+            <item index="26" class="java.lang.String" itemvalue="distlib" />
+            <item index="27" class="java.lang.String" itemvalue="pytz" />
+            <item index="28" class="java.lang.String" itemvalue="itchat" />
+            <item index="29" class="java.lang.String" itemvalue="idna" />
+            <item index="30" class="java.lang.String" itemvalue="Pillow" />
+          </list>
+        </value>
+      </option>
+    </inspection_tool>
+    <inspection_tool class="PyPep8NamingInspection" enabled="true" level="WEAK WARNING" enabled_by_default="true">
+      <option name="ignoredErrors">
+        <list>
+          <option value="N802" />
+          <option value="N803" />
+          <option value="N806" />
+        </list>
+      </option>
+    </inspection_tool>
+  </profile>
+</component>
