+++ output_file/2019-6-4/station_geo.txt
@@ -1,0 +1,31 @@
+苹果园 116.374375,39.907325
+古城 116.190337,39.907450
+八角游乐园 116.212684,39.907442
+国贸 116.461841,39.909104
+四惠 116.493449,39.909048
+西苑 116.291278,39.997796
+国家图书馆 116.325190,39.943114
+黄村西大街 116.332631,39.731769
+黄村火车站 116.332611,39.722966
+天坛东门 116.420833,39.882558
+天通苑南 116.412661,40.066458
+天通苑北 116.417635,39.977146
+花园桥 116.444183,39.940486
+金台路 116.478065,39.915614
+郝家府 116.460904,39.922979
+东夏园 116.732371,39.902924
+虎坊桥 116.384596,39.889486
+百子湾 116.497768,39.892530
+安华桥 116.394655,39.968507
+丰台南路 116.296748,39.840444
+安贞门 116.405954,39.977005
+农业展览馆 116.465145,39.940528
+西局 116.444183,39.940486
+张郭庄 116.187193,39.858100
+九龙山 116.478695,39.893222
+崔各庄 116.492968,40.022201
+双桥 116.581181,39.912359
+生命科学园 116.340917,39.976391
+良乡大学城西 116.156282,39.723157
+星城 116.060902,39.713463
+大石河东 116.041483,39.710345
